+++ src/frontend/apps/lti_site/apps/.cookiecutter/{{cookiecutter.app_name}}/utils/parseDataElements/parseDataElements.ts
@@ -4,6 +4,7 @@ import { {{cookiecutter.model}}ModelName as modelName } from 'lib-components';
 export const parseDataElements = (element: Element): {{cookiecutter.app_name_capitalized}}AppData => {
   const context = JSON.parse(element.getAttribute('data-context')!);
 
+  context.resource_id = context.resource?.id;
   if (context.modelName === modelName.{{cookiecutter.model_plural_capitalized}}) {
     context.{{cookiecutter.model_lower}} = context.resource;
     delete context.resource;