+++ src/frontend/apps/lti_site/apps/deposit/utils/parseDataElements/parseDataElements.ts
@@ -4,6 +4,7 @@ import { FileDepositoryModelName as modelName } from 'lib-components';
 export const parseDataElements = (element: Element): DepositAppData => {
   const context = JSON.parse(element.getAttribute('data-context')!);
 
+  context.resource_id = context.resource?.id;
   if (context.modelName === modelName.FileDepositories) {
     context.fileDepository = context.resource;
     delete context.resource;