+++ src/frontend/apps/lti_site/utils/parseDataElements/parseDataElements.ts
@@ -2,7 +2,7 @@ import { AppData, modelName } from 'lib-components';
 
 export const parseDataElements = (element: Element): AppData => {
   const context = JSON.parse(element.getAttribute('data-context')!);
-
+  context.resource_id = context.resource?.id;
   if (context.modelName) {
     switch (context.modelName) {
       case modelName.VIDEOS: